requirements/base.txt
boto==2.43.0
Django==1.10.2
django-ckeditor==5.1.1
django-countries==4.0
django-solo==1.1.2
django-storages-redux==1.3.2
django-widget-tweaks==1.4.1
easy_thumbnails==2.3
psycopg2==2.6.2
python-slugify==1.2.1
six==1.10.0

